+++ pkg/streamingpromql/operators/binops/binary_operation.go
@@ -5,10 +5,13 @@ package binops
 import (
 	"errors"
 	"fmt"
+	"maps"
 	"math"
 	"slices"
+	"strings"
 	"time"
 
+	"github.com/grafana/regexp"
 	"github.com/prometheus/prometheus/model/histogram"
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/model/timestamp"
@@ -745,4 +748,43 @@ var boolComparisonOperationFuncs = map[parser.ItemType]binaryOperationFunc{
 
 		return 0, nil, true, true, nil
 	},
+}
+
+type QueryHints struct {
+	Include []string
+}
+
+func BuildMatchers(metadata []types.SeriesMetadata, hints *QueryHints) types.Matchers {
+	var matchers []*labels.Matcher
+
+	for _, field := range hints.Include {
+		values := make(map[string]struct{})
+
+		for _, series := range metadata {
+			val := series.Labels.Get(field)
+			if val != "" {
+				values[regexp.QuoteMeta(val)] = struct{}{}
+			}
+		}
+
+		// TODO: Skip using this particular field if there are more than N values. Note
+		//  that since we're building a regex, we need to not include it at all if we exceed
+		//  the max number of values rather than include only a few matchers.
+
+		if len(values) != 0 {
+			ordered := make([]string, 0, len(values))
+			for k := range maps.Keys(values) {
+				ordered = append(ordered, k)
+			}
+
+			slices.Sort(ordered)
+			matchers = append(matchers, &labels.Matcher{
+				Type:  labels.MatchRegexp,
+				Name:  field,
+				Value: strings.Join(ordered, "|"),
+			})
+		}
+	}
+
+	return matchers
 }