+++ pkg/streamingpromql/optimize/plan/commonsubexpressionelimination/range_vector_operator.go
@@ -59,11 +59,11 @@ func (b *RangeVectorDuplicationBuffer) AddConsumer() *RangeVectorDuplicationCons
 	}
 }
 
-func (b *RangeVectorDuplicationBuffer) SeriesMetadata(ctx context.Context) ([]types.SeriesMetadata, error) {
+func (b *RangeVectorDuplicationBuffer) SeriesMetadata(ctx context.Context, matchers types.Matchers) ([]types.SeriesMetadata, error) {
 	if b.seriesMetadataCount == 0 {
 		// Haven't loaded series metadata yet, load it now.
 		var err error
-		b.seriesMetadata, err = b.Inner.SeriesMetadata(ctx)
+		b.seriesMetadata, err = b.Inner.SeriesMetadata(ctx, matchers)
 		if err != nil {
 			return nil, err
 		}
@@ -324,8 +324,8 @@ type RangeVectorDuplicationConsumer struct {
 
 var _ types.RangeVectorOperator = &RangeVectorDuplicationConsumer{}
 
-func (d *RangeVectorDuplicationConsumer) SeriesMetadata(ctx context.Context) ([]types.SeriesMetadata, error) {
-	return d.Buffer.SeriesMetadata(ctx)
+func (d *RangeVectorDuplicationConsumer) SeriesMetadata(ctx context.Context, matchers types.Matchers) ([]types.SeriesMetadata, error) {
+	return d.Buffer.SeriesMetadata(ctx, matchers)
 }
 
 func (d *RangeVectorDuplicationConsumer) NextSeries(ctx context.Context) error {